final_ex/aggregator/requirements.txt
pika
loguru
pandas
cassandra-driver
matplotlib
seaborn
